Initial work on splitting out BitcoindRpcClient into multiple files
